save
又上了一天课，我服了呀！！

+++ src/common.h
@@ -42,7 +42,7 @@
     //正常情况下在硬件上应该是光栅化应该是直接向pixel shader 传送当前点的属性，并且针对这些属性直接进行渲染，但是事实上这么做并不方便，所以现在的
     //做法是光栅化对一个三角形完全执行完，把这些属性值存储在容器里。在光栅化结束之后将数据从容器送入pixel shader
     struct RasterToPixel{
-        Eigen::Vector2f index ;
+        Eigen::Vector2i index ;
         Eigen::Vector3f Word3DPoint ;
         Eigen::Vector3f Normal ;
         Eigen::Vector2f TexCoord ;
